+++ release_notes.md
@@ -40,6 +40,7 @@ Description......
 * `findFirstRow` renamed to `indexOfFirstMatchingRow` on `TDBQuery`.
 * `findFirstRowFromIndex:` renamed to `indexOfFirstMatchingRowFromIndex:` on `TDBQuery`.
 * Return `NSNotFound` when appropriate.
+* `addRow` and `insertRow` do not return any value.
 
 ### Enhancements:
 * `renameColumnWithIndex:to:` has been added to `TDBTable`.